Organize and fix tests
