+++ activemq-nob-supervisor/src/test/java/org/apache/activemq/nob/supervisor/SupervisorTest.java
@@ -1,0 +1,17 @@
+/**
+ */
+package org.apache.activemq.nob.supervisor;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+public class SupervisorTest {
+
+    @Test
+    public void testXbeanTemplate() throws Exception {
+        String template = SupervisorService.getXbeanConfigurationTemplate();
+        Assert.assertNotNull(template);
+        Assert.assertTrue(template.indexOf("\"$brokerName\"") > 0);
+    }
+
+}
